requirements.txt
Django==2.0
django_debug_toolbar==1.9.1
djangorestframework==3.7.3
argon2_cffi==16.3.0
mysqlclient==1.3.12
